Refactored the "service-user" lab to test for a real-world example of user and group permissions
This lab should now test for WordPress's ability to auto-update, and separately for the presence of a new user with read/write permissions on WordPress's directory.

This should be more tangible for learners, and give them a good starting point for doing their own research into how to solve this problem.

This will require a rethink on how to mark this, and I would also invite discussion on what to rename this lab to.